+++ build_docs.py
@@ -1,5 +1,4 @@
 import os
-import sys
 import subprocess
 
 import warp as wp
@@ -12,9 +11,8 @@ wp.init()
 # disable sphinx color output
 os.environ["NO_COLOR"] = "1"
 
-function_ref = open("docs/modules/functions.rst","w")
-wp.print_builtins(function_ref)
-function_ref.close()
+with open("docs/modules/functions.rst","w") as function_ref:
+    wp.print_builtins(function_ref)
 
 # run Sphinx build
 try: