+++ warp/codegen.py
@@ -93,17 +93,25 @@ def _get_struct_instance_ctype(
                 assert value.dtype == var_type.dtype, "assign to struct member variable {} failed, expected type {}, got type {}".format(field_name, var_type.dtype, value.dtype)
                 setattr(inst_ctype, field_name, value.__ctype__())
         elif isinstance(var_type, Struct):
-            _get_struct_instance_ctype(value, inst_ctype, field_name)
+            if value is None:
+                _get_struct_instance_ctype(StructInstance(var_type), inst_ctype, field_name)
+            else:
+                _get_struct_instance_ctype(value, inst_ctype, field_name)
         elif issubclass(var_type, ctypes.Array):
             # vector/matrix type, e.g. vec3
-            if types_equal(type(value), var_type):
+            if value is None:
+                setattr(inst_ctype, field_name, var_type())
+            elif types_equal(type(value), var_type):
                 setattr(inst_ctype, field_name, value)
             else:
                 # conversion from list/tuple, ndarray, etc.
-                setattr(inst_ctype, field_name, var_type(*value))
+                setattr(inst_ctype, field_name, var_type(value))
         else:
             # primitive type
-            setattr(inst_ctype, field_name, var_type._type_(value))
+            if value is None:
+                setattr(inst_ctype, field_name, var_type._type_())
+            else:
+                setattr(inst_ctype, field_name, var_type._type_(value))
 
     return inst_ctype
 
@@ -120,7 +128,7 @@ def _fmt_struct_instance_repr(inst: StructInstance, depth: int) -> str:
         if field_name == "_dummy_":
             continue
 
-        field_value = getattr(inst, field_name)
+        field_value = getattr(inst, field_name, None)
 
         if isinstance(field_value, StructInstance):
             field_value = _fmt_struct_instance_repr(field_value, depth + 1)
@@ -1801,18 +1809,36 @@ WP_API void {name}_cpu_backward({reverse_args});
 # converts a constant Python value to equivalent C-repr
 def constant_str(value):
     
-    if type(value) == bool:
+    value_type = type(value)
+
+    if value_type == bool:
         if value:
             return "true"
         else:
             return "false"
 
-    elif type(value) == str:
+    elif value_type == str:
         # ensure constant strings are correctly escaped
         return "\"" + str(value.encode("unicode-escape").decode()) + "\""
 
     elif isinstance(value, ctypes.Array):
-        return "{" + ", ".join(map(str, value)) + "}"
+        if value_type._wp_scalar_type_ == float16:
+            # special case for float16, which is stored as uint16 in the ctypes.Array
+            from warp.context import runtime
+            scalar_value = runtime.core.half_bits_to_float
+        else:
+            scalar_value = lambda x: x
+
+        # list of scalar initializer values
+        initlist = []
+        for i in range(value._length_):
+            x = ctypes.Array.__getitem__(value, i)
+            initlist.append(str(scalar_value(x)))
+
+        dtypestr = f"wp::initializer_array<{value._length_},wp::{value._wp_scalar_type_.__name__}>"
+
+        # construct value from initializer array, e.g. wp::initializer_array<4,wp::float32>{1.0, 2.0, 3.0, 4.0}
+        return f"{dtypestr}{{{', '.join(initlist)}}}"
 
     else:
         # otherwise just convert constant to string